requirements.txt
Django==1.7.4
SQLAlchemy==0.6.9
argparse==1.2.1
cjklib==0.3.2
django-bootstrap3==4.4.1
django-crispy-forms==1.4.0
django-selectable==0.8.0
django-timedeltafield==0.7.0
greenlet==0.4.1
meld3==0.6.10
stevedore==0.11
wsgiref==0.1.2
tornado==4.1
